// WADNR.Web/src/app/shared/generated/model/service-forestry-upload-result.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ServiceForestryUploadResult { 
    RecordsImported?: number;
    ElapsedSeconds?: number;
    Warnings?: Array<string> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ServiceForestryUploadResultForm { 
    RecordsImported?: FormControl<number>;
    ElapsedSeconds?: FormControl<number>;
    Warnings?: FormControl<Array<string>>;
}

export class ServiceForestryUploadResultFormControls { 
    public static RecordsImported = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ElapsedSeconds = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Warnings = (value: FormControlState<Array<string>> | Array<string> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<string>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}