// WADNR.Web/src/app/shared/generated/model/tabular-data-import-grid-row.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TabularDataImportGridRow { 
    TabularDataImportID?: number;
    TabularDataImportTableTypeID?: number;
    UploadDate?: string | null;
    UploadPersonName?: string | null;
    LastProcessedDate?: string | null;
    LastProcessedPersonName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TabularDataImportGridRowForm { 
    TabularDataImportID?: FormControl<number>;
    TabularDataImportTableTypeID?: FormControl<number>;
    UploadDate?: FormControl<string>;
    UploadPersonName?: FormControl<string>;
    LastProcessedDate?: FormControl<string>;
    LastProcessedPersonName?: FormControl<string>;
}

export class TabularDataImportGridRowFormControls { 
    public static TabularDataImportID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TabularDataImportTableTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static UploadDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static UploadPersonName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LastProcessedDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LastProcessedPersonName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}